requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-pptx==0.6.21 